Implement type checking, GCed strings
Include sds (https://github.com/antirez/sds) set up to use libgc
(the Boehm GC) for memory allocation.

Types are checked at the level of the intermediate representation
for the arguments of operators and functions and for the result
variable of SET. If a variable is undeclared its type is inferred
to be the resulting type of the IR expression it is first assigned.
Functions can be declared without an implementation using the
keyword "DECLARE" before "SUB". This feature is useful for C
interoperability. Functions and operators have a single return
type (no overloading), with "AND", "OR" and "XOR" being hard-coded
exceptions (they act as either numerical or boolean depending on
the arguments). Casts/type conversions are not yet implemented.
For now, any integral type can be used with any other provided
they are either both signed or both unsigned.